NN-1632 removing options from viewportal to allow user to change zoom and splitting out Sizing into component and index js files
